Enum hyperion_vfs::tree::Node

pub enum Node<Mut> {
    File(FileRef<Mut>),
    Directory(DirRef<Mut>),
}

Variants

File(FileRef<Mut>)

a normal file, like /etc/fstab

or

a device mapped to a file, like /dev/fb0

Directory(DirRef<Mut>)

a directory with 0 or more files, like /home/

or

a device mapped to a directory, like /https/archlinux/org/

mountpoints are also directory devices

directory devices may be unlistable, because it’s not sensible to list every website there is

a directory device most likely contains more directory devices, like /https/archlinux/org inside /https/archlinux/

Implementations

impl<Mut: AnyMutex> Node<Mut>

pub fn new_root() -> Self

pub fn find(&self, path: impl AsRef<Path>, make_dirs: bool) -> IoResult<Self>

pub fn find_dir( &self, path: impl AsRef<Path>, make_dirs: bool ) -> IoResult<DirRef<Mut>>

pub fn find_file( &self, path: impl AsRef<Path>, make_dirs: bool, create: bool ) -> IoResult<FileRef<Mut>>

pub fn install_dev_with( &self, path: impl AsRef<Path>, dev: impl FileDevice + 'static )

pub fn insert_file( &self, path: impl AsRef<Path>, make_dirs: bool, dev: FileRef<Mut> ) -> IoResult<()>

pub fn insert_dir( &self, path: impl AsRef<Path>, make_dirs: bool, dev: DirRef<Mut> ) -> IoResult<()>

pub fn insert( &self, path: impl AsRef<Path>, make_dirs: bool, node: Node<Mut> ) -> IoResult<()>

pub fn install_dev( &self, path: impl AsRef<Path>, dev: impl FileDevice + 'static )

pub fn install_dev_ref(&self, path: impl AsRef<Path>, dev: FileRef<Mut>)